+++ lib/src/pages/dialog/state.dart
@@ -1,0 +1,4 @@
+import 'package:app_state/app_state.dart';
+import 'package:flutter/widgets.dart';
+
+class QuickDialogNotifier<R> extends ChangeNotifier with PageStateMixin<R> {}