+++ lib/src/widgets/auth/firebase_userpic.dart
@@ -1,5 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+
+import 'userpic.dart';
 
 class FirebaseUserpicWidget extends StatelessWidget {
   final double size;
@@ -13,15 +15,9 @@ class FirebaseUserpicWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final url = user.photoURL;
-
-    return DecoratedBox(
-      decoration: const BoxDecoration(
-        shape: BoxShape.circle,
-      ),
-      child: url == null
-          ? Icon(Icons.person, size: size)
-          : Image.network(url, width: size, height: size),
+    return UserpicWidget(
+      user.photoURL,
+      size: size,
     );
   }
 }